Add solver for linear systems.
The solver does implement the Gauß-Seidel method of iteratively solving
linear equations.  This commit also moves the burden of handling control
points away from the PointOfLine implementations and towards classes derived
from Layout.  The idea is that it should be possible to add new layouts
without having to modify the structure of the graph representation.